+++ dependencies/src/main/java/com/google/cloud/tools/opensource/dependencies/DependencyGraph.java
@@ -16,6 +16,7 @@
 
 package com.google.cloud.tools.opensource.dependencies;
 
+import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -25,6 +26,7 @@ import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Queue;
 import java.util.Set;
 
 import org.eclipse.aether.artifact.Artifact;
@@ -49,6 +51,18 @@ import com.google.common.collect.TreeMultimap;
  */
 public class DependencyGraph {
 
+  private static final class LevelOrderQueueItem {
+    final DependencyNode dependencyNode;
+  
+    // Null for the first item
+    final DependencyPath parentPath;
+  
+    LevelOrderQueueItem(DependencyNode dependencyNode, DependencyPath parentPath) {
+      this.dependencyNode = dependencyNode;
+      this.parentPath = parentPath;
+    }
+  }
+
   // DependencyGraphBuilder builds this in breadth first order, unless explicitly stated otherwise.
   // That is, this list contains the paths to each node in breadth first order 
   private final List<DependencyPath> graph = new ArrayList<>();
@@ -65,9 +79,6 @@ public class DependencyGraph {
   private SetMultimap<String, DependencyPath> paths = HashMultimap.create();
 
   private DependencyNode root;
-  
-  DependencyGraph() {
-  }
 
   public DependencyGraph(DependencyNode root) {
     this.root = root;
@@ -207,16 +218,72 @@ public class DependencyGraph {
     return ImmutableList.copyOf(visitor.getPaths());
   }
 
-  public void setRoot(DependencyNode root) {
-    this.root = root;
-  }
-
   private final Set<Artifact> checkedArtifacts = new HashSet<>();
   
   void addUnresolvableArtifactProblem(Artifact artifact) {
     if (checkedArtifacts.add(artifact)) {
       artifactProblems.add(createUnresolvableArtifactProblem(artifact));
     }
+  }
+
+  /**
+   * Builds a dependency graph by traversing dependency tree in level-order (breadth-first search).
+   *
+   * @param root node to start traversal
+   */
+  public static DependencyGraph from(DependencyNode root) {
+    DependencyGraph graph = new DependencyGraph(root);
+  
+    return levelOrder(graph);
+  }
+
+  // a bit weird; this modifies the argument and returns it
+  static DependencyGraph levelOrder(DependencyGraph graph) {
+    Queue<DependencyGraph.LevelOrderQueueItem> queue = new ArrayDeque<>();
+    queue.add(new DependencyGraph.LevelOrderQueueItem(graph.root, null));
+  
+    while (!queue.isEmpty()) {
+      DependencyGraph.LevelOrderQueueItem item = queue.poll();
+      DependencyNode dependencyNode = item.dependencyNode;
+  
+      DependencyPath parentPath = item.parentPath;
+      Artifact artifact = dependencyNode.getArtifact();
+      if (artifact != null && parentPath != null) {
+        // When requesting dependencies of 2 or more artifacts, root DependencyNode's artifact is
+        // set to null
+  
+        // When there's an ancestor dependency node with the same groupId and artifactId as
+        // the dependency, Maven will not pick up the dependency. For example, if there's a
+        // dependency path "g1:a1:2.0 / ... / g1:a1:1.0" (the leftmost node as root), then Maven's
+        // dependency mediation always picks g1:a1:2.0 over g1:a1:1.0.
+        
+        // TODO This comment doesn't seem right. That's true for the root,
+        // but not for non-root nodes. A node elsewhere in the tree could cause the 
+        // descendant to be selected. 
+        
+        String groupIdAndArtifactId = Artifacts.makeKey(artifact);
+        boolean ancestorHasSameKey =
+            parentPath.getArtifacts().stream()
+                .map(Artifacts::makeKey)
+                .anyMatch(key -> key.equals(groupIdAndArtifactId));
+        if (ancestorHasSameKey) {
+          continue;
+        }
+      }
+  
+      // parentPath is null for the first item
+      DependencyPath path =
+          parentPath == null
+              ? new DependencyPath(artifact)
+              : parentPath.append(dependencyNode.getDependency());
+      graph.addPath(path);
+  
+      for (DependencyNode child : dependencyNode.getChildren()) {
+        queue.add(new DependencyGraph.LevelOrderQueueItem(child, path));
+      }
+    }
+  
+    return graph;
   }
   
 }
